edit testing and datatype for phone number

+++ server/database/build.sql
@@ -12,7 +12,7 @@ CREATE TABLE users (
   name VARCHAR(100) NOT NULL,
   email VARCHAR(255) NOT NULL UNIQUE,
   password VARCHAR(255) NOT NULL,
-  phone INT NOT NULL,
+  phone BIGINT NOT NULL,
   location TEXT NOT NULL
 );
